# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  phone: string;
  password: string;
  countryCode: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface UserProfile {
  id: number;
  phone: string;
  countryCode: string;
  balance: number;
  totalEarned: number;
  referralCode: string;
  referralLink?: string;
  vipLevel: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UserStats {
  balance: number;
  totalEarned: number;
  totalWithdrawn: number;
  dailyEarnings: number;
  activeProducts: number;
  totalInvited: number;
  totalRewards: number;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  durationDays: number;
  dailyReturn: number;
  totalReturn: number;
  imageUrl: string;
  category: string;
  description: string;
  featured?: boolean;
}

export interface UserProduct {
  id: number;
  productId: number;
  productName: string;
  purchaseDate: string;
  expiryDate: string;
  dailyReturn: number;
  totalReturn: number;
  earnedSoFar: number;
  status: string;
  imageUrl: string;
}

export interface MissionLevel {
  level: number;
  name: string;
  description: string;
  commission: number;
  required: number;
  current: number;
  objective: number;
  rewards: number;
  status: string;
}

export interface MissionsResponse {
  totalRewards: number;
  totalCount: number;
  levels: MissionLevel[];
}

export interface VipLevel {
  commission: number;
  users: number;
  rewards: number;
}

export interface TeamResponse {
  referralCode: string;
  referralLink: string;
  totalInvited: number;
  totalRewards: number;
  vip1: VipLevel;
  vip2: VipLevel;
  vip3: VipLevel;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  status: string;
  createdAt: string;
  description: string;
}

export interface RechargeInput {
  amount: number;
  method: string;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  proofUrl?: string | null;
}

export interface WithdrawInput {
  amount: number;
  bankAccountId: number;
}

export interface CheckinStatus {
  checkedInToday: boolean;
  streak: number;
  totalRewards: number;
  dailyReward: number;
  /** @nullable */
  nextCheckinAt: string | null;
}

export interface CheckinResult {
  reward: number;
  streak: number;
  totalRewards: number;
  message: string;
}

export interface BankAccount {
  id: number;
  bankName: string;
  accountNumber: string;
  accountHolder: string;
  isDefault: boolean;
  createdAt: string;
}

export interface BankAccountInput {
  bankName: string;
  accountNumber: string;
  accountHolder: string;
  isDefault?: boolean;
}