# lib/api-zod/src/generated/types/bankAccountInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BankAccountInput {
  bankName: string;
  accountNumber: string;
  accountHolder: string;
  isDefault?: boolean;
}