# lib/api-zod/src/generated/types/checkinStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinStatus {
  checkedInToday: boolean;
  streak: number;
  totalRewards: number;
  dailyReward: number;
  /** @nullable */
  nextCheckinAt: string | null;
}