# lib/api-zod/src/generated/types/missionsResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MissionLevel } from './missionLevel';

export interface MissionsResponse {
  totalRewards: number;
  totalCount: number;
  levels: MissionLevel[];
}