# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  price: number;
  durationDays: number;
  dailyReturn: number;
  totalReturn: number;
  imageUrl: string;
  category: string;
  description: string;
  featured?: boolean;
}