# lib/api-zod/src/generated/types/rechargeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RechargeInput {
  amount: number;
  method: string;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  proofUrl?: string | null;
}