# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  phone: string;
  password: string;
  countryCode: string;
  /** @nullable */
  referralCode?: string | null;
}