# lib/api-zod/src/generated/types/teamResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VipLevel } from './vipLevel';

export interface TeamResponse {
  referralCode: string;
  referralLink: string;
  totalInvited: number;
  totalRewards: number;
  vip1: VipLevel;
  vip2: VipLevel;
  vip3: VipLevel;
}