# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * McDonald's Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  phone: string;
  countryCode: string;
  balance: number;
  totalEarned: number;
  referralCode: string;
  referralLink?: string;
  vipLevel: number;
  createdAt: string;
}